+++ keys/generate.sh
@@ -12,7 +12,7 @@ temp_subj="/C=DE/ST=Some state/L=Some city/O=Some organization/OU=Some departmen
 root_subj="$temp_subj"
 
 subj_suffix_arr=( "/CN=dtool-looktup-server" "/CN=token-generator-ldap" "/CN=mongodb" "/CN=ldap" )
-subdir_arr=( "dtool_lookup_server" "token_generator_ldap" "mongodb" "ldap" )
+subdir_arr=( "dserver" "token_generator_ldap" "mongodb" "ldap" )
 
 PASSW=$(openssl rand -base64 32)
 echo "$PASSW" > passw