+++ keys/generate_jwt_key.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
-mkdir -p dtool_lookup_server
+mkdir -p dserver
 mkdir -p token_generator_ldap
 
-openssl genrsa -out dtool_lookup_server/jwt_key 2048
-openssl rsa -in dtool_lookup_server/jwt_key -pubout -outform PEM -out dtool_lookup_server/jwt_key.pub
+openssl genrsa -out dserver/jwt_key 2048
+openssl rsa -in dserver/jwt_key -pubout -outform PEM -out dserver/jwt_key.pub
 
-cp dtool_lookup_server/jwt_key token_generator_ldap/jwt_key
-cp dtool_lookup_server/jwt_key.pub token_generator_ldap/jwt_key.pub
+cp dserver/jwt_key token_generator_ldap/jwt_key
+cp dserver/jwt_key.pub token_generator_ldap/jwt_key.pub